requirements.txt
lml>=0.0.6
pyecharts-jupyter-installer==0.0.3
